[Refactor] #1 : Sourse, Resource 폴더 구조 추가 및 ListModel dummyData 분리
