+++ HeckDesigns/Source/Components/FavoriteSampleView.swift
